Even more simplifications
MCM tests now all start from an OpenQASM program.
