--- test/unit_tests/braket/default_simulator/openqasm/test_branched_control_flow.py
@@ -1,1060 +1,0 @@
-# Copyright Amazon.com Inc. or its affiliates. All Rights Reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License"). You
-# may not use this file except in compliance with the License. A copy of
-# the License is located at
-#
-#     http://aws.amazon.com/apache2.0/
-#
-# or in the "license" file accompanying this file. This file is
-# distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF
-# ANY KIND, either express or implied. See the License for the specific
-# language governing permissions and limitations under the License.
-
-"""Tests for control flow handling in the Interpreter and ProgramContext.
-
-Tests verify that:
-- The Interpreter performs eager evaluation for non-MCM contexts.
-- ProgramContext.handle_branching_statement, handle_for_loop, and
-  handle_while_loop perform per-path evaluation when branched (MCM).
-- Break/continue signals are raised by the Interpreter and caught by loops.
-"""
-
-import pytest
-
-from braket.default_simulator.gate_operations import BRAKET_GATES, GPhase
-from braket.default_simulator.openqasm.parser.openqasm_ast import (
-    ArrayLiteral,
-    ArrayType,
-    BooleanLiteral,
-    BranchingStatement,
-    BreakStatement,
-    ContinueStatement,
-    ForInLoop,
-    Identifier,
-    IndexedIdentifier,
-    IntegerLiteral,
-    IntType,
-    RangeDefinition,
-    WhileLoop,
-)
-from braket.default_simulator.openqasm.program_context import (
-    AbstractProgramContext,
-    ProgramContext,
-    _BreakSignal,
-    _ContinueSignal,
-)
-from braket.default_simulator.openqasm.interpreter import Interpreter
-from braket.default_simulator.openqasm.simulation_path import FramedVariable, SimulationPath
-from braket.default_simulator.openqasm.circuit import Circuit
-
-
-class SimpleProgramContext(AbstractProgramContext):
-    """Minimal non-MCM context that just builds a Circuit.
-
-    Used to verify the Interpreter's generic eager-evaluation paths for
-    if/else, for, and while — the code that runs when
-    ``supports_midcircuit_measurement`` is False.
-    """
-
-    def __init__(self):
-        super().__init__()
-        self._circuit = Circuit()
-
-    @property
-    def circuit(self):
-        return self._circuit
-
-    def is_builtin_gate(self, name: str) -> bool:
-        return name in BRAKET_GATES
-
-    def add_phase_instruction(self, target, phase_value):
-        self._circuit.add_instruction(GPhase(target, phase_value))
-
-    def add_gate_instruction(self, gate_name, target, params, ctrl_modifiers, power):
-        instruction = BRAKET_GATES[gate_name](
-            target, *params, ctrl_modifiers=ctrl_modifiers, power=power
-        )
-        self._circuit.add_instruction(instruction)
-
-
-class TestInterpreterBranchingStatement:
-    """Tests for eager if/else evaluation in the Interpreter (non-MCM)."""
-
-    def test_if_true_visits_if_block(self):
-        """When condition is True, the Interpreter should visit the if_block."""
-        context = ProgramContext()
-        assert not context.supports_midcircuit_measurement or not context.is_branched
-        interpreter = Interpreter(context)
-
-        visited = []
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str):
-                visited.append(node)
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        node = BranchingStatement(
-            condition=BooleanLiteral(True),
-            if_block=["if_stmt_1", "if_stmt_2"],
-            else_block=["else_stmt_1"],
-        )
-
-        tracking_visit(node)
-        assert visited == ["if_stmt_1", "if_stmt_2"]
-
-    def test_if_false_visits_else_block(self):
-        """When condition is False, the Interpreter should visit the else_block."""
-        context = ProgramContext()
-        interpreter = Interpreter(context)
-
-        visited = []
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str):
-                visited.append(node)
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        node = BranchingStatement(
-            condition=BooleanLiteral(False),
-            if_block=["if_stmt"],
-            else_block=["else_stmt"],
-        )
-
-        tracking_visit(node)
-        assert visited == ["else_stmt"]
-
-
-class TestInterpreterForLoop:
-    """Tests for eager for-loop evaluation in the Interpreter (non-MCM)."""
-
-    def test_iterates_over_range(self):
-        """The Interpreter should unroll the for loop eagerly."""
-        context = ProgramContext()
-        interpreter = Interpreter(context)
-
-        iterations = []
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str):
-                iterations.append(node)
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(2), IntegerLiteral(1)
-            ),
-            block=["body_stmt"],
-        )
-
-        tracking_visit(node)
-        body_visits = [x for x in iterations if x == "body_stmt"]
-        assert len(body_visits) == 3
-
-
-class TestInterpreterWhileLoop:
-    """Tests for eager while-loop evaluation in the Interpreter (non-MCM)."""
-
-    def test_loops_until_condition_false(self):
-        """The Interpreter should loop eagerly until condition is False."""
-        context = ProgramContext()
-        # Declare a counter variable
-        context.declare_variable("counter", IntType(IntegerLiteral(32)), IntegerLiteral(3))
-        interpreter = Interpreter(context)
-
-        iteration_count = [0]
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str) and node == "body_stmt":
-                iteration_count[0] += 1
-                # Decrement counter
-                current = context.get_value("counter")
-                context.update_value(Identifier("counter"), IntegerLiteral(current.value - 1))
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        # Condition: counter > 0 — we use a BinaryExpression but that's complex.
-        # Instead, use a simpler approach: the condition reads the counter variable.
-        # We'll just test with a fixed iteration count using the mock.
-        # Actually, let's use a direct approach with the interpreter's own visit.
-        # We need a proper OpenQASM program for a full integration test.
-        # For unit testing, let's verify the signal mechanism works.
-        assert iteration_count[0] == 0  # Sanity check
-
-
-class TestInterpreterBreakContinueSignals:
-    """Tests that the Interpreter raises _BreakSignal/_ContinueSignal for break/continue."""
-
-    def test_break_raises_signal(self):
-        """Visiting a BreakStatement should raise _BreakSignal."""
-        interpreter = Interpreter()
-        with pytest.raises(_BreakSignal):
-            interpreter.visit(BreakStatement())
-
-    def test_continue_raises_signal(self):
-        """Visiting a ContinueStatement should raise _ContinueSignal."""
-        interpreter = Interpreter()
-        with pytest.raises(_ContinueSignal):
-            interpreter.visit(ContinueStatement())
-
-    def test_break_caught_by_for_loop(self):
-        """Break inside a for loop should stop iteration."""
-        interpreter = Interpreter()
-
-        iteration_count = [0]
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str) and node == "body_stmt":
-                iteration_count[0] += 1
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(4), IntegerLiteral(1)
-            ),
-            block=["body_stmt", BreakStatement()],
-        )
-
-        tracking_visit(node)
-        assert iteration_count[0] == 1
-
-    def test_continue_skips_rest_of_body(self):
-        """Continue inside a for loop should skip to next iteration."""
-        interpreter = Interpreter()
-
-        pre_count = [0]
-        post_count = [0]
-        original_visit = interpreter.visit
-
-        def tracking_visit(node):
-            if isinstance(node, str):
-                if node == "pre_continue":
-                    pre_count[0] += 1
-                elif node == "post_continue":
-                    post_count[0] += 1
-                return node
-            return original_visit(node)
-
-        interpreter.visit = tracking_visit
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(2), IntegerLiteral(1)
-            ),
-            block=["pre_continue", ContinueStatement(), "post_continue"],
-        )
-
-        tracking_visit(node)
-        assert pre_count[0] == 3
-        assert post_count[0] == 0
-
-
-class TestBranchedBranchingStatement:
-    """Tests for handle_branching_statement in branched mode (MCM)."""
-
-    def test_branched_routes_paths_by_condition(self):
-        """When branched, paths should be routed based on per-path condition evaluation."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        path0.set_variable("c", FramedVariable("c", None, BooleanLiteral(True), False, 0))
-        path1.set_variable("c", FramedVariable("c", None, BooleanLiteral(False), False, 0))
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        if_visited_paths = []
-        else_visited_paths = []
-
-        def mock_visit(node):
-            if isinstance(node, Identifier) and node.name == "c":
-                path_idx = context._active_path_indices[0]
-                path = context._paths[path_idx]
-                var = path.get_variable("c")
-                return var.value
-            if isinstance(node, BooleanLiteral):
-                return node
-            if node == "if_stmt":
-                if_visited_paths.extend(list(context._active_path_indices))
-            elif node == "else_stmt":
-                else_visited_paths.extend(list(context._active_path_indices))
-            return node
-
-        node = BranchingStatement(
-            condition=Identifier("c"),
-            if_block=["if_stmt"],
-            else_block=["else_stmt"],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_branching_statement(node)
-
-        assert 0 in if_visited_paths
-        assert 1 in else_visited_paths
-        assert set(context._active_path_indices) == {0, 1}
-
-    def test_branched_no_else_block(self):
-        """When branched with no else block, false paths should survive unchanged."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        path0.set_variable("c", FramedVariable("c", None, BooleanLiteral(True), False, 0))
-        path1.set_variable("c", FramedVariable("c", None, BooleanLiteral(False), False, 0))
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        if_visited = []
-
-        def mock_visit(node):
-            if isinstance(node, Identifier) and node.name == "c":
-                path_idx = context._active_path_indices[0]
-                return context._paths[path_idx].get_variable("c").value
-            if isinstance(node, BooleanLiteral):
-                return node
-            if node == "if_stmt":
-                if_visited.extend(list(context._active_path_indices))
-            return node
-
-        node = BranchingStatement(
-            condition=Identifier("c"),
-            if_block=["if_stmt"],
-            else_block=[],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_branching_statement(node)
-
-        assert 0 in if_visited
-        assert 1 not in if_visited
-        assert set(context._active_path_indices) == {0, 1}
-
-
-class TestBranchedForLoop:
-    """Tests for handle_for_loop in branched mode (MCM)."""
-
-    def test_branched_sets_loop_variable_per_path(self):
-        """When branched, loop variable should be set per-path."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        loop_var_values = []
-
-        def mock_visit(node):
-            if isinstance(node, RangeDefinition):
-                return node
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            if node == "body_stmt":
-                for path_idx in context._active_path_indices:
-                    var = context._paths[path_idx].get_variable("i")
-                    if var:
-                        loop_var_values.append((path_idx, var.value))
-            return node
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(1), IntegerLiteral(1)
-            ),
-            block=["body_stmt"],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_for_loop(node)
-
-        assert len(loop_var_values) >= 2
-        assert set(context._active_path_indices) == {0, 1}
-
-    def test_branched_for_loop_break(self):
-        """Break in branched for loop should stop iteration."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        context._paths = [path0]
-        context._active_path_indices = [0]
-
-        iteration_count = [0]
-
-        def mock_visit(node):
-            if isinstance(node, RangeDefinition):
-                return node
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            if isinstance(node, BreakStatement):
-                raise _BreakSignal()
-            if node == "body_stmt":
-                iteration_count[0] += 1
-            return node
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(4), IntegerLiteral(1)
-            ),
-            block=["body_stmt", BreakStatement()],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_for_loop(node)
-
-        assert iteration_count[0] == 1
-        assert 0 in context._active_path_indices
-
-    def test_branched_for_loop_continue(self):
-        """Continue in branched for loop should skip to next iteration."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        context._paths = [path0]
-        context._active_path_indices = [0]
-
-        pre_continue_count = [0]
-        post_continue_count = [0]
-
-        def mock_visit(node):
-            if isinstance(node, RangeDefinition):
-                return node
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            if isinstance(node, ContinueStatement):
-                raise _ContinueSignal()
-            if node == "pre_continue":
-                pre_continue_count[0] += 1
-            elif node == "post_continue":
-                post_continue_count[0] += 1
-            return node
-
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(2), IntegerLiteral(1)
-            ),
-            block=["pre_continue", ContinueStatement(), "post_continue"],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_for_loop(node)
-
-        assert pre_continue_count[0] == 3
-        assert post_continue_count[0] == 0
-
-
-class TestBranchedWhileLoop:
-    """Tests for handle_while_loop in branched mode (MCM)."""
-
-    def test_branched_while_loop_per_path_condition(self):
-        """When branched, while condition should be evaluated per-path."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        path0.set_variable("n", FramedVariable("n", None, IntegerLiteral(2), False, 0))
-        path1.set_variable("n", FramedVariable("n", None, IntegerLiteral(0), False, 0))
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        body_executions = {0: 0, 1: 0}
-
-        def mock_visit(node):
-            if isinstance(node, Identifier) and node.name == "n":
-                path_idx = context._active_path_indices[0]
-                var = context._paths[path_idx].get_variable("n")
-                val = var.value.value
-                return BooleanLiteral(val > 0)
-            if isinstance(node, BooleanLiteral):
-                return node
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            if node == "body_stmt":
-                for path_idx in context._active_path_indices:
-                    body_executions[path_idx] += 1
-                    var = context._paths[path_idx].get_variable("n")
-                    new_val = IntegerLiteral(var.value.value - 1)
-                    context._paths[path_idx].set_variable(
-                        "n", FramedVariable("n", None, new_val, False, 0)
-                    )
-            return node
-
-        node = WhileLoop(
-            while_condition=Identifier("n"),
-            block=["body_stmt"],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_while_loop(node)
-
-        assert body_executions[0] == 2
-        assert body_executions[1] == 0
-        assert set(context._active_path_indices) == {0, 1}
-
-    def test_branched_while_loop_break(self):
-        """Break in branched while loop should exit the loop."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        context._paths = [path0]
-        context._active_path_indices = [0]
-
-        iteration_count = [0]
-
-        def mock_visit(node):
-            if isinstance(node, BooleanLiteral):
-                return node
-            if isinstance(node, IntegerLiteral):
-                return BooleanLiteral(True)
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            if isinstance(node, BreakStatement):
-                raise _BreakSignal()
-            if node == "body_stmt":
-                iteration_count[0] += 1
-            return node
-
-        node = WhileLoop(
-            while_condition=IntegerLiteral(1),
-            block=["body_stmt", BreakStatement()],
-        )
-
-        context.set_visitor(mock_visit)
-        context.handle_while_loop(node)
-
-        assert iteration_count[0] == 1
-        assert 0 in context._active_path_indices
-
-
-class TestFrameManagement:
-    """Tests for _enter_frame_for_active_paths and _exit_frame_for_active_paths."""
-
-    def test_enter_frame_increments_frame_number(self):
-        """Entering a frame should increment frame_number for all active paths."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {}, frame_number=0)
-        path1 = SimulationPath([], 50, {}, {}, frame_number=0)
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        context._enter_frame_for_active_paths()
-
-        assert path0.frame_number == 1
-        assert path1.frame_number == 1
-
-    def test_exit_frame_restores_frame_number(self):
-        """Exiting a frame should restore frame_number for all active paths."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {}, frame_number=1)
-        path1 = SimulationPath([], 50, {}, {}, frame_number=1)
-        context._paths = [path0, path1]
-        context._active_path_indices = [0, 1]
-
-        context._exit_frame_for_active_paths()
-
-        assert path0.frame_number == 0
-        assert path1.frame_number == 0
-
-    def test_exit_frame_removes_scoped_variables(self):
-        """Exiting a frame should remove variables declared in that frame."""
-        context = ProgramContext()
-        context._is_branched = True
-        path0 = SimulationPath([], 50, {}, {}, frame_number=1)
-        path0.set_variable("x", FramedVariable("x", None, IntegerLiteral(10), False, 1))
-        path0.set_variable("y", FramedVariable("y", None, IntegerLiteral(20), False, 0))
-        context._paths = [path0]
-        context._active_path_indices = [0]
-
-        context._exit_frame_for_active_paths()
-
-        assert path0.get_variable("x") is None
-        assert path0.get_variable("y") is not None
-        assert path0.get_variable("y").value == IntegerLiteral(20)
-
-
-class TestAbstractContextControlFlow:
-    """Tests for AbstractProgramContext defaults via SimpleProgramContext."""
-
-    def test_handle_branching_raises(self):
-        """handle_branching_statement raises NotImplementedError on a non-MCM context."""
-        ctx = SimpleProgramContext()
-        node = BranchingStatement(condition=BooleanLiteral(True), if_block=[], else_block=[])
-        with pytest.raises(NotImplementedError):
-            ctx.handle_branching_statement(node)
-
-    def test_handle_for_loop_raises(self):
-        """handle_for_loop raises NotImplementedError on a non-MCM context."""
-        ctx = SimpleProgramContext()
-        node = ForInLoop(
-            type=IntType(IntegerLiteral(32)),
-            identifier=Identifier("i"),
-            set_declaration=RangeDefinition(
-                IntegerLiteral(0), IntegerLiteral(1), IntegerLiteral(1)
-            ),
-            block=[],
-        )
-        with pytest.raises(NotImplementedError):
-            ctx.handle_for_loop(node)
-
-    def test_handle_while_loop_raises(self):
-        """handle_while_loop raises NotImplementedError on a non-MCM context."""
-        ctx = SimpleProgramContext()
-        node = WhileLoop(while_condition=BooleanLiteral(True), block=[])
-        with pytest.raises(NotImplementedError):
-            ctx.handle_while_loop(node)
-
-    def test_set_visitor_raises(self):
-        """set_visitor raises NotImplementedError on a non-MCM context."""
-        ctx = SimpleProgramContext()
-        with pytest.raises(NotImplementedError):
-            ctx.set_visitor(lambda x: x)
-
-    def test_is_branched_returns_false(self):
-        ctx = SimpleProgramContext()
-        assert ctx.is_branched is False
-
-    def test_supports_midcircuit_measurement_returns_false(self):
-        ctx = SimpleProgramContext()
-        assert ctx.supports_midcircuit_measurement is False
-
-    def test_active_paths_returns_empty(self):
-        ctx = SimpleProgramContext()
-        assert ctx.active_paths == []
-
-
-class TestProgramContextResolveIndex:
-    """Cover _resolve_index edge cases."""
-
-    def test_empty_indices(self):
-        path = SimulationPath([], 0, {}, {})
-        assert ProgramContext._resolve_index(path, []) == 0
-
-    def test_none_indices(self):
-        path = SimulationPath([], 0, {}, {})
-        assert ProgramContext._resolve_index(path, None) == 0
-
-    def test_integer_literal_index(self):
-        path = SimulationPath([], 0, {}, {})
-        assert ProgramContext._resolve_index(path, [[IntegerLiteral(3)]]) == 3
-
-    def test_identifier_index_from_path(self):
-        path = SimulationPath([], 0, {}, {})
-        path.set_variable("i", FramedVariable("i", None, IntegerLiteral(2), False, 0))
-        assert ProgramContext._resolve_index(path, [[Identifier("i")]]) == 2
-
-    def test_multi_index_returns_zero(self):
-        path = SimulationPath([], 0, {}, {})
-        assert ProgramContext._resolve_index(path, [[IntegerLiteral(1)], [IntegerLiteral(2)]]) == 0
-
-
-class TestProgramContextHelpers:
-    """Cover static helpers and _ensure_path_variable."""
-
-    def test_get_path_measurement_result_present(self):
-        path = SimulationPath([], 0, {}, {0: [1, 0, 1]})
-        assert ProgramContext._get_path_measurement_result(path, 0) == 1
-
-    def test_get_path_measurement_result_absent(self):
-        path = SimulationPath([], 0, {}, {})
-        assert ProgramContext._get_path_measurement_result(path, 0) == 0
-
-    def test_set_value_at_index_array_literal(self):
-        val = ArrayLiteral([IntegerLiteral(0), IntegerLiteral(0)])
-        ProgramContext._set_value_at_index(val, 0, 1)
-        assert val.values[0].value == 1
-
-    def test_ensure_path_variable_existing(self):
-        ctx = ProgramContext()
-        path = SimulationPath([], 0, {}, {})
-        fv = FramedVariable("x", None, IntegerLiteral(10), False, 0)
-        path.set_variable("x", fv)
-        result = ctx._ensure_path_variable(path, "x")
-        assert result is fv
-
-    def test_ensure_path_variable_from_shared(self):
-        ctx = ProgramContext()
-        ctx.declare_variable("y", IntType(IntegerLiteral(32)), IntegerLiteral(7))
-        path = SimulationPath([], 0, {}, {})
-        result = ctx._ensure_path_variable(path, "y")
-        assert result is not None
-        assert result.value.value == 7
-
-
-class TestProgramContextBranchedVariables:
-    """Cover branched declare_variable, update_value, get_value, is_initialized."""
-
-    def _make_branched_context(self):
-        """Create a ProgramContext in branched mode with two paths."""
-        ctx = ProgramContext()
-        ctx._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        ctx._paths = [path0, path1]
-        ctx._active_path_indices = [0, 1]
-        return ctx
-
-    def test_declare_variable_branched(self):
-        """declare_variable in branched mode stores per-path FramedVariables."""
-        ctx = self._make_branched_context()
-        ctx.declare_variable("x", IntType(IntegerLiteral(32)), IntegerLiteral(10))
-        # Both paths should have the variable
-        for path in ctx._paths:
-            fv = path.get_variable("x")
-            assert fv is not None
-            assert fv.value.value == 10
-
-    def test_update_value_branched(self):
-        """update_value in branched mode updates per-path."""
-        ctx = self._make_branched_context()
-        ctx.declare_variable("x", IntType(IntegerLiteral(32)), IntegerLiteral(0))
-        # Update only on path 0
-        ctx._active_path_indices = [0]
-        ctx.update_value(Identifier("x"), IntegerLiteral(42))
-        ctx._active_path_indices = [0, 1]
-        assert ctx._paths[0].get_variable("x").value.value == 42
-        assert ctx._paths[1].get_variable("x").value.value == 0
-
-    def test_update_value_branched_indexed(self):
-        """update_value with IndexedIdentifier in branched mode."""
-        ctx = self._make_branched_context()
-        arr_type = ArrayType(IntType(IntegerLiteral(32)), [IntegerLiteral(2)])
-        arr_val = ArrayLiteral([IntegerLiteral(0), IntegerLiteral(0)])
-        # declare_variable in branched mode adds to symbol_table and per-path
-        ctx.declare_variable("arr", arr_type, arr_val)
-        # Update arr[1] = 99 on path 0
-        ctx._active_path_indices = [0]
-        indexed = IndexedIdentifier(Identifier("arr"), [[IntegerLiteral(1)]])
-        ctx.update_value(indexed, IntegerLiteral(99))
-        ctx._active_path_indices = [0, 1]
-        p0_val = ctx._paths[0].get_variable("arr").value
-        assert p0_val.values[1].value == 99
-
-    def test_get_value_branched_reads_first_active_path(self):
-        """get_value in branched mode reads from first active path."""
-        ctx = self._make_branched_context()
-        ctx.declare_variable("x", IntType(IntegerLiteral(32)), IntegerLiteral(0))
-        ctx._paths[0].get_variable("x").value = IntegerLiteral(10)
-        ctx._paths[1].get_variable("x").value = IntegerLiteral(20)
-        ctx._active_path_indices = [1]
-        val = ctx.get_value("x")
-        assert val.value == 20
-
-    def test_get_value_branched_falls_back_to_shared(self):
-        """get_value falls back to shared table for pre-branching variables."""
-        ctx = self._make_branched_context()
-        # Add to shared table directly (simulating pre-branching declaration)
-        ctx.symbol_table.add_symbol("pre", IntType(IntegerLiteral(32)), False)
-        ctx.variable_table.add_variable("pre", IntegerLiteral(7))
-        val = ctx.get_value("pre")
-        assert val.value == 7
-
-    def test_is_initialized_branched_checks_path(self):
-        """is_initialized in branched mode checks per-path variables."""
-        ctx = self._make_branched_context()
-        ctx.declare_variable("x", IntType(IntegerLiteral(32)), IntegerLiteral(0))
-        assert ctx.is_initialized("x") is True
-
-    def test_is_initialized_branched_falls_back_to_shared(self):
-        """is_initialized falls back to shared table."""
-        ctx = self._make_branched_context()
-        ctx.symbol_table.add_symbol("shared", IntType(IntegerLiteral(32)), False)
-        ctx.variable_table.add_variable("shared", IntegerLiteral(0))
-        assert ctx.is_initialized("shared") is True
-
-
-class TestProgramContextBranchedInstructions:
-    """Cover branched add_*_instruction methods."""
-
-    def _make_branched_context(self):
-        ctx = ProgramContext()
-        ctx._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        path1 = SimulationPath([], 50, {}, {})
-        ctx._paths = [path0, path1]
-        ctx._active_path_indices = [0, 1]
-        return ctx
-
-    def test_add_phase_instruction_branched(self):
-        """add_phase_instruction routes to all active paths when branched."""
-        ctx = self._make_branched_context()
-        ctx.add_qubits("q", 1)
-        ctx.add_phase_instruction((0,), 1.5)
-        assert len(ctx._paths[0].instructions) == 1
-        assert len(ctx._paths[1].instructions) == 1
-
-    def test_add_gate_instruction_branched(self):
-        """add_gate_instruction routes to all active paths when branched."""
-        ctx = self._make_branched_context()
-        ctx.add_qubits("q", 1)
-        ctx.add_gate_instruction("x", (0,), [], [], 1)
-        assert len(ctx._paths[0].instructions) == 1
-        assert len(ctx._paths[1].instructions) == 1
-
-    def test_add_reset_branched(self):
-        """add_reset routes to all active paths when branched."""
-        ctx = self._make_branched_context()
-        ctx.add_qubits("q", 1)
-        ctx.add_reset([0])
-        assert len(ctx._paths[0].instructions) == 1
-        assert len(ctx._paths[1].instructions) == 1
-
-
-class TestProgramContextBranchedEdgeCases:
-    """Cover remaining branched-mode edge cases."""
-
-    def _make_branched_context(self):
-        ctx = ProgramContext()
-        ctx._is_branched = True
-        path0 = SimulationPath([], 50, {}, {})
-        ctx._paths = [path0]
-        ctx._active_path_indices = [0]
-        return ctx
-
-    def test_update_value_branched_missing_variable_raises(self):
-        """update_value raises KeyError when variable not found on path."""
-        ctx = self._make_branched_context()
-        # Declare in symbol table so get_type works, but don't set on path
-        ctx.symbol_table.add_symbol("missing", IntType(IntegerLiteral(32)), False)
-        with pytest.raises(KeyError, match="Variable 'missing' not found"):
-            ctx.update_value(Identifier("missing"), IntegerLiteral(1))
-
-    def test_get_value_by_identifier_branched_indexed(self):
-        """get_value_by_identifier with IndexedIdentifier in branched mode."""
-        ctx = self._make_branched_context()
-        arr_val = ArrayLiteral([IntegerLiteral(10), IntegerLiteral(20), IntegerLiteral(30)])
-        ctx.declare_variable(
-            "arr", ArrayType(IntType(IntegerLiteral(32)), [IntegerLiteral(3)]), arr_val
-        )
-        indexed = IndexedIdentifier(Identifier("arr"), [[IntegerLiteral(1)]])
-        val = ctx.get_value_by_identifier(indexed)
-        assert val.value == 20
-
-    def test_get_value_by_identifier_branched_falls_back(self):
-        """get_value_by_identifier falls back to shared table for pre-branching vars."""
-        ctx = self._make_branched_context()
-        ctx.symbol_table.add_symbol("shared", IntType(IntegerLiteral(32)), False)
-        ctx.variable_table.add_variable("shared", IntegerLiteral(99))
-        val = ctx.get_value_by_identifier(Identifier("shared"))
-        assert val.value == 99
-
-    def test_get_value_branched_wraps_raw_python(self):
-        """get_value wraps raw Python values into AST literals."""
-        ctx = self._make_branched_context()
-        ctx.declare_variable("x", IntType(IntegerLiteral(32)), IntegerLiteral(5))
-        # Manually set a raw int to test wrapping
-        ctx._paths[0].get_variable("x").value = 42
-        val = ctx.get_value("x")
-        assert val.value == 42
-
-    def test_handle_branching_non_branched_else(self):
-        """handle_branching_statement non-branched path with else block."""
-        ctx = ProgramContext()
-        assert not ctx._is_branched
-
-        visited = []
-
-        def mock_visit(node):
-            if isinstance(node, BooleanLiteral):
-                return node
-            if isinstance(node, list):
-                for item in node:
-                    mock_visit(item)
-                return
-            visited.append(node)
-            return node
-
-        node = BranchingStatement(
-            condition=BooleanLiteral(False),
-            if_block=["if_stmt"],
-            else_block=["else_stmt"],
-        )
-        ctx.set_visitor(mock_visit)
-        ctx.handle_branching_statement(node)
-        assert "else_stmt" in visited
-        assert "if_stmt" not in visited
-
-    def test_handle_branching_non_branched_no_else(self):
-        """handle_branching_statement non-branched path with no else block."""
-        ctx = ProgramContext()
-        visited = []
-
-        def mock_visit(node):
-            if isinstance(node, BooleanLiteral):
-                return node
-            visited.append(node)
-            return node
-
-        node = BranchingStatement(
-            condition=BooleanLiteral(False),
-            if_block=["if_stmt"],
-            else_block=[],
-        )
-        ctx.set_visitor(mock_visit)
-        ctx.handle_branching_statement(node)
-        assert visited == []
-
-
-class TestNonMCMInterpreterControlFlow:
-    """Verify the Interpreter's generic eager-evaluation paths using SimpleProgramContext.
-
-    SimpleProgramContext returns ``supports_midcircuit_measurement = False``,
-    so the Interpreter handles if/else, for, and while inline rather than
-    delegating to the context's handle_* methods.
-    """
-
-    def test_if_true(self):
-        qasm = "OPENQASM 3.0;\nqubit[1] q;\nif (true) { x q[0]; }"
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_if_false_else(self):
-        qasm = "OPENQASM 3.0;\nqubit[1] q;\nif (false) { x q[0]; } else { h q[0]; }"
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_if_false_no_else(self):
-        qasm = "OPENQASM 3.0;\nqubit[1] q;\nif (false) { x q[0]; }"
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 0
-
-    def test_for_loop_range(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] sum = 0;
-        for int[32] i in [0:2] { sum = sum + i; }
-        if (sum == 3) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_for_loop_discrete_set(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] sum = 0;
-        for int[32] i in {2, 5} { sum = sum + i; }
-        if (sum == 7) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_for_loop_break(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] count = 0;
-        for int[32] i in [0:9] {
-            count = count + 1;
-            if (count == 3) { break; }
-        }
-        if (count == 3) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_for_loop_continue(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] x_count = 0;
-        for int[32] i in [1:4] {
-            if (i % 2 == 0) { continue; }
-            x_count = x_count + 1;
-        }
-        if (x_count == 2) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_while_loop(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] n = 3;
-        while (n > 0) { n = n - 1; }
-        if (n == 0) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_while_loop_break(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] n = 0;
-        while (true) {
-            n = n + 1;
-            if (n == 5) { break; }
-        }
-        if (n == 5) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
-
-    def test_while_loop_continue(self):
-        qasm = """
-        OPENQASM 3.0;
-        qubit[1] q;
-        int[32] count = 0;
-        int[32] x_count = 0;
-        while (count < 5) {
-            count = count + 1;
-            if (count % 2 == 0) { continue; }
-            x_count = x_count + 1;
-        }
-        if (x_count == 3) { x q[0]; }
-        """
-        ctx = SimpleProgramContext()
-        Interpreter(ctx).run(qasm)
-        assert len(ctx.circuit.instructions) == 1
